# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/phone/one-click": {
      "post": {
        "operationId": "AuthController_oneClickLogin",
        "summary": "一键登录（阿里云）",
        "description": "通过阿里云一键登录SDK获取的accessToken进行一键登录，后端自动获取手机号并注册/登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OneClickLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/LoginResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/phone/send-code": {
      "post": {
        "operationId": "AuthController_sendSmsCode",
        "summary": "发送短信验证码",
        "description": "向指定手机号发送短信验证码",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendSmsCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "boolean"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "发送过于频繁，请稍后再试",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/phone/code-login": {
      "post": {
        "operationId": "AuthController_phoneCodeLogin",
        "summary": "验证码登录",
        "description": "通过手机号和短信验证码进行登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhoneCodeLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/LoginResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误或验证码错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "验证码发送过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "summary": "刷新访问令牌",
        "description": "使用有效的JWT令牌刷新获取新的访问令牌",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/RefreshTokenResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "令牌无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/third-party/douyin": {
      "post": {
        "operationId": "AuthController_douyinLogin",
        "summary": "抖音登录",
        "description": "通过抖音授权token进行登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DouyinLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/LoginResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "抖音token验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/third-party/apple": {
      "post": {
        "operationId": "AuthController_appleLogin",
        "summary": "Apple登录",
        "description": "通过Apple ID token进行登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/LoginResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Apple token验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/user": {
      "get": {
        "operationId": "UserController_getUser",
        "summary": "获取用户基本信息",
        "description": "获取当前登录用户的基本信息",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserProfileResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "post": {
        "operationId": "UserController_postUser",
        "summary": "创建用户",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutUserInfoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserProfileResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      },
      "put": {
        "operationId": "UserController_editUser",
        "summary": "修改用户",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutUserInfoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserProfileResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
...
    "/api/agreement/user": {
      "get": {
        "operationId": "AgreementController_getUserAgreement",
        "summary": "获取用户协议",
        "description": "获取用户协议内容",
        "parameters": [
          {
            "name": "language",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgreementResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "agreement"
        ]
      }
    },
    "/api/agreement/privacy": {
      "get": {
        "operationId": "AgreementController_getPrivacyPolicy",
        "summary": "获取隐私政策",
        "description": "获取隐私政策内容",
        "parameters": [
          {
            "name": "language",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgreementResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "agreement"
        ]
      }
    },
    "/api/agreement/mobile-auth": {
      "get": {
        "operationId": "AgreementController_getMobileAuthTerms",
        "summary": "获取移动认证服务条款",
        "description": "获取中国移动认证服务条款内容",
        "parameters": [
          {
            "name": "language",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgreementResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "agreement"
        ]
      }
    },
...
  },
  "info": {
    "title": "Claw Step API",
    "description": "Claw Step 应用后端 API 文档",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "ResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "状态码",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "提示信息",
            "example": "success"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "访问令牌",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "description": "刷新令牌",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "isNew": {
            "type": "boolean",
            "description": "是否为新用户",
            "example": false
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "isNew"
        ]
      },
      "OneClickLoginDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "移动商SDK返回的accessToken"
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "状态码",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "错误信息",
            "example": "请求参数错误"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "SendSmsCodeDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "description": "手机号",
            "example": "13800138000"
          }
        },
        "required": [
          "phone"
        ]
      },
      "PhoneCodeLoginDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "description": "手机号",
            "example": "13800138000"
          },
          "code": {
            "type": "string",
            "description": "短信验证码",
            "example": "123456"
          }
        },
        "required": [
          "phone",
          "code"
        ]
      },
      "RefreshTokenResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "新的访问令牌",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "description": "新的刷新令牌",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "刷新令牌",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "DouyinLoginDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "抖音授权accessToken",
            "example": "douyin_access_token_here"
          },
          "openid": {
            "type": "string",
            "description": "抖音用户ID（openid）",
            "example": "douyin_openid_here"
          },
          "nickname": {
            "type": "string",
            "description": "抖音用户昵称",
            "example": "抖音用户"
          },
          "avatar": {
            "type": "string",
            "description": "抖音用户头像URL",
            "example": "https://example.com/avatar.jpg"
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "AppleLoginDto": {
        "type": "object",
        "properties": {
          "identityToken": {
            "type": "string",
            "description": "Apple ID token (JWT)",
            "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "userIdentifier": {
            "type": "string",
            "description": "Apple用户标识（user identifier）",
            "example": "apple_user_id_here"
          },
          "email": {
            "type": "string",
            "description": "Apple用户邮箱",
            "example": "user@example.com"
          },
          "fullName": {
            "type": "string",
            "description": "Apple用户全名",
            "example": "John Doe"
          }
        },
        "required": [
          "identityToken"
        ]
      },
      "UserProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户ID",
            "example": "uuid-string"
          },
          "displayId": {
            "type": "string",
            "description": "用户ID（显示用）",
            "example": "43455483"
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "username"
          },
          "nickname": {
            "type": "string",
            "description": "昵称",
            "example": "假装看风景"
          },
          "avatar": {
            "type": "string",
            "description": "头像URL",
            "example": "https://example.com/avatar.jpg"
          },
          "isPro": {
            "type": "boolean",
            "description": "是否为Pro用户",
            "example": true
          },
          "travelTickets": {
            "type": "number",
            "description": "旅行券数量",
            "example": 3
          },
          "language": {
            "type": "string",
            "description": "语言设置",
            "example": "zh_CN"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "phone": {
            "type": "string",
            "description": "手机号"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          }
        },
        "required": [
          "id",
          "displayId",
          "nickname",
          "isPro",
          "travelTickets",
          "language",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "PutUserInfoDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "avatar": {
            "type": "string",
            "description": "用户头像"
          },
          "nickname": {
            "type": "string",
            "description": "昵称"
          },
          "language": {
            "type": "string",
            "description": "语言"
          }
        },
        "required": [
          "username",
          "avatar",
          "nickname",
          "language"
        ]
      },
...
      "AgreementResponseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "协议标题",
            "example": "用户协议"
          },
          "content": {
            "type": "string",
            "description": "协议内容",
            "example": "欢迎您使用..."
          },
          "language": {
            "type": "string",
            "description": "语言代码",
            "example": "zh_CN"
          },
          "version": {
            "type": "string",
            "description": "版本号",
            "example": "1.0"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "title",
          "content",
          "language",
          "version",
          "updatedAt"
        ]
      },
...
    }
  }
}